+++ helpers/project.py
@@ -3,20 +3,12 @@ from scrapy.utils.conf import get_config
 from scrapycw.helpers import Helper
 
 
-class ProjectListHelper(Helper):
+class ProjectHelper(Helper):
 
-    def get_value(self):
+    def list(self):
         config = get_config()
-        return [project for project, _ in config.items('settings')]
-
-    def get_json(self):
-        projects = []
-        for project in self.get_value():
-            projects.append({
-                "name": project
-            })
         return {
             "success": True,
             "message": None,
-            "projects": projects,
+            "projects": [project for project, _ in config.items('settings')],
         }