+++ web/api/urls.py
@@ -6,4 +6,5 @@ urlpatterns = [
     path('project-list', views.project_list, name='project-list'),
     path('spider-list', views.spider_list, name='spider-list'),
     path('crawl', views.crawl, name='crawl'),
+    path('stop', views.stop, name='stop'),
 ]